requirements.txt
aiofiles==24.1.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
Authlib==1.5.2
bcrypt==4.3.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
dateparser==1.2.1
dotenv==0.9.9
ecdsa==0.19.1
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cache2==0.2.2
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
hf-xet==1.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.31.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
llvmlite==0.44.0
lru-dict==1.3.0
lupa==2.4
lxml==5.4.0
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mediawiki_langcodes==0.2.15
more-itertools==10.6.0
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.2.4
openai-whisper==20240930
orjson==3.11.2
packaging==25.0
passlib==1.7.4
pendulum==3.1.0
pgvector==0.4.1
pillow==11.2.1
preshed==3.0.9
psutil==7.0.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.4.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
starlette==0.46.1
sympy==1.13.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
wikitextprocessor @ git+https://github.com/tatuylonen/wikitextprocessor.git@ea19a0a7c5b481aa68246c6ccff1c86eb8f8be26
wiktextract @ git+https://github.com/tatuylonen/wiktextract.git@30661e8eb5bfcb9de03322ca17d007056ee0204f
wrapt==1.17.2
